Add TEST-LIBREBOL native
This provides a place and switches for building libRebol API tests into
the core.  It would be more desirable to not have them in the core but
to actually go through a build process with a separate executable.
But that would require the tests finding the toolchain, using CALL to
run the compiler, using CALL to run the executables, and connecting a
lot of the dots for that...which would be difficult cross-platform.

Rather than put off having a place to put such tests any longer, this
just makes an optional TEST-LIBREBOL native that debug builds can
include.  Then the tests call that.  It returns a BLOCK! which is
pairs of the test number and a true or false status on the success.

This commit only adds the ability while currently thinking of it (so
there's only one test of the testing system).  But by having it there,
any bugs found in libRebol can now have tests included with the commit
that fixes them.