Workaround JS-NATIVE registration inside rebPromise()
Whenever libRebol invokes the body of a JS-NATIVE, that body needs to
be run on the main / GUI thread.  Otherwise, it would be of limited
use--since it couldn't interact with the DOM.  Hence when Rebol code
inside a rebPromise() was running on a worker thread, it would have
to bounce the request for execution over to the main thread and wait
on that to be signaled with mutexes/conditions.

This worked.  However, one thing that didn't take the need to bounce
into account was JS-NATIVE itself.  It would register a function in
a table, but if that registration happened during a promise...the
entry to the table would be on the worker thread.  Hence it would not
be visible to the bounced code.

This commit contains a workaround that sneakily avoids adding
thread-aware complexity to JS-NATIVE.  Instead, it makes JS-NATIVE
itself a JavaScript function...which then calls through to the actual
code to add the native (JS-NATIVE-MAINTHREAD).  It knows that when
it does so it will be on the main thread, because that's the guarantee
it's given by the system.

A more efficient version of this same idea would just factor the logic
differently inside the JavaScript module.  But it was interesting to
do it this way just to test that it *could* work, because it seemed
like it should!