Make PARSE COPY give just BLOCK!/GROUP!/TEXT!/BINARY!
PARSE's COPY command was changed between Rebol2 and R3-Alpha to
preserve the BINARY!/string distinction.  Previously a parsed binary
would come back as a string:

    rebol2>> parse #{42434445} [copy x to end]
    == true

    >> x
    == "BCDE"

This was changed--but beyond changing just this case, it was assumed
that the returned type from COPY should match the type of the input.
This meant that copying a fragment out of a URL! would give a URL!
(for instance) even though it was not well-formed.

    >> parse http://example.com ["http://" thru "." copy x to end]
    == true

    >> x
    == com

    >> type? x
    == url!

Other awkward situations arose for copying pieces out of PATH!, which
are now considered illegal in Ren-C due to path immutability and
making them follow specific rules (they are not generic containers like
BLOCK! and GROUP! are):

    r3-alpha>> parse 'a/b/c [skip skip copy x to end]
    == true

    r3-alpha>> x
    == c

    r3-alpha>> type? x
    == path!

This commit keeps the choice to preserve BLOCK! and GROUP! distinction,
as well as the BINARY! distinction, but makes PARSE of a string always
COPY a TEXT!.  This helps even in cases like tags, where copying `<bc>`
out of `<abcd>` may not make an illegal tag, but makes a somewhat
unnatural suggestion that `<b` or `c>` were in the initial sequence.
The neturality of text in printing no delimiters avoids that.