Make it possible to AWAIT in JS-AWAITER
This adds the ability to use the JavaScript AWAIT statement in a
JS-AWAITER.  It does so by labeling those functions `async`.

These functions were previously not expected to return a value, rather
to (eventually) call one of the resolve() or reject() functions passed
to them as implicit parameters.  With this change, they will return
an ES6 Promise.  If AWAIT is called then that promise will not be
in a fulfilled state when the promise is returned, and it must be
waited on.