First attempt at making reject() / throw() not crash JS
In adding handling for the reject() / throw() case, this embraces an
asymmetry where rebPromise()'s .then() handler receives a REBVAL*, but
the .catch() always receives a JavaScript value (typically a JavaScript
Error Object, which tries to embed knowledge of the stack trace of the
origin, though that is not enforced in the language).

This is an incomplete first try at getting %mod-javascript.c to follow
the convention.  It is to the point of successfully avoiding crashes
in the emterpreted and pthreads version when a throw() occurs inside
a JS-AWAITER.  The idea it uses is that a reference to what JavaScript
raised is held in a table.  Then a number to find that is put into
a Rebol-styled THROW.  If the Rebol throw makes it up the stack to
an API that doesn't have a JavaScript-catch attached to it, then it
emerges as a Rebol error.

The approach is flawed and leaks numbers, and the necessary mechanics
for translating objects to Rebol errors is punted on (instead making
a placeholder string-based "user error").  But this is closer than it
was, and is a baseline for further investigation.