Add DEBUG_PRINTF_FAIL_LOCATIONS
It's a difficult balance to debug code built with emscripten.  Ren-C's
debug build is very aggressive by default, and builds prohibitively
large files...and many emscripten debug settings can't be used on
such codebases without crashing the emcc compiler.

So debug switches have to be fairly selective to helping find the most
relevant bugs specific to JavaScript.  (Any bug applicable to a non web
build as well is better debugged in address-sanitized and heavily
symbolized C)

This adds a small switch that just expands the fail() macro to say the
file and line in the C source where the fail originated from.  This
helps specifically in the debugging of how Rebol failures are turned
into JavaScript objects for .catch() clauses, and vice-versa.