Fix CELL_FLAG_UNEVALUATED bugs, add strengthening assert
The UNEVALUATED flag is meant solely as a flag for letting functions
tell the difference between things like `if [x > 1] [...]` and
`data: [x > 1] | if data [...]`  The idea is that sometimes this
knowledge is useful, such as in that case for catching new users making
a classic mistake of thinking that IF requires its condition in a
BLOCK! the way that WHILE does.

As an experimental flag, it has shown a lot of value in accomplishing
compatibility detections when an interface on a function changes.
It is not a tool to reach for in changing the behavior of a function,
but it is good for raising errors in a "linting" style.

It's a bit complex to maintain it correctly, and the least expensive
way to maintain it in places it needs to be cleared is just to let it
be implicitly cleared by any Move_Value(), Derelativize() or
Init_Value()).  Yet usages have shown cases where the bit is
incorrectly set from time to time; and as it's not a bit that's checked
very often these cases are hard to notice.

This fixes a case where ACTION! results were not having the flag
cleared on their results, which would rarely come up as a problem
unless the action used low level evaluative services.  SWITCH was
doing so, and what makes the most sense is just to clear the bit when
the output value is received on all actions.

It also adds a debug check that tries to synchronize what is known
about the inertness of the datatype in the switch() with what the
result can possibly be.  This should improve the accuracy of the bit.