Add reb.Binary() on ArrayBuffer for JavaScript
Because the original API definition was biased to C, the rebBinary()
creation function didn't mention that it took a size in its interface.
However, the API in JavaScript ES6 for dealing with memory has
dedicated array buffer and view types, which is used whenever it is
working with binary data.  It has no size.

This adds support for a relatively efficient way of creating a binary
out of an arraybuffer that does not require setting up an intermediate
heap buffer to pass to the (now arity-2) rebSizedBinary() routine.