Repurpose SET/ANY and GET/ANY to VOID! handling
Introduction of the NULL state in Ren-C was an important philosophical
development, facilitating a large number of interesting features.
Initially it was seen as being roughly analogous to what had been an
UNSET!, but without the ability to be put in blocks...because it
conveyed the complete absence of a value.

This led to flux on naming and meaning of what SET of a NULL would do,
if it should error or not, and what the refinement should be called.
/ANY was considered a somewhat poor name.

Over time the responsibilities shifted, so that the duties of what had
been UNSET! and NONE! were split between three things: NULLs, VOID!s,
and BLANK!s.  This new landscape implied more tolerance of NULLs,
seeing them as being more similar to a non-valued version of NONE!
than to how UNSET! had behaved.  And VOID! took on a role that didn't
have anything specifically to do with being unset.

This restores the /ANY refinements to GET and SET to relate to the
handling of VOID!.  The meanings are changed rather significantly from
older Rebol, but the usages wind up being interestingly mostly
compatible.  While voids aren't what unset variables hold (they hold
*no* value, best communicated with NULL) the situations that reify
NULLs do often make voids.  And one of the big sources of historical
unsets--functions that don't return values like PRINT--now return
VOID! values.