Move byte-order sensitive macros to %sys-rebnod.h
The motivations for the byte-order macros are entirely for the Rebol
Node.  Now that there's a %sys-rebnod.h, it makes sense to put them
in that file--which helps provide reasoning for their existence, vs.
having to make that argument in the abstract in %reb-c.h

Also corrects an improperly arranged commenting situation regarding
the use of stdbool.h