Streamline creation of rebFooQ and rebFoo API variants
In order to have new API entry points, there generally had to be a
function in the %a-lib.c file.  As a first cut at creating the "Q"
splice quoting variants of every API, they were done manually and
then set up to call a common function.

Such a boilerplate task is more suited to the %make-reb-lib.r code
to generate.  So this commit does that.  It reduces the overall amount
of code and makes things clearer.

Also, it ties up a loose end on the handling of functions which are
variadic but have additional parameters besides the variadic ones.
The JavaScript still doesn't handle those, but most of them were
related to getting back additional parameters for multiple return
values...and JavaScript callers wouldn't have heap addresses to pass
to those anyway.