Use macros naming slots in PARSE subframe call
When a recursion into PARSE via SUBPARSE was happening, it was using
raw integer indices to set up the slots.  Seemingly it used the wrong
numbers.  This changes it to use the same macros naming the slots as
the parse code itself does.

Also removes a faulty assert from MATCH