Avoid assert on uninitialized first feed value fetch
An assertion for fetching the next value in the feed checks that it is
not an END marker.  This assert happened on all Fetch_Next_In_Feed()
even if it was the first.

For some reason most compilations didn't notice this use of an
uninitialized value in an inline function (debug build only).  However,
a RIGOROUS C89 build on a certain linux version seemed to notice.

This creates a differentiated Fetch_First_In_Feed() entry point that
does not perform the assert.