Use casts instead of AS_REBCHR(), lose AS_REBYTE_PTR()
It's desirable to make it clear when a REBYTE* is transitioning to a
"guaranteed valid UTF-8 sequence" via a REBCHR(*).  This makes that
expression done with a cast operator instead of the more noisy and
obscure AS_REBCHR() macro.

The other direction--of going from a REBCHR(*) to a REBYTE* is probably
the least of one's concerns.  It might arguably be nice to know one is
losing the "guaranteed UTF-8 status" and have a point in the source to
see that.  But you'd find out because of the inability to implicitly
convert back to REBCHR() anyway.