Encoded UTF-8 representation of char in payload, updates
This commit started from an examination of uses of GET_ANY_CHAR() to
see which ones could be eliminated.  Ultimately it wound up adding
the UTF-8 representation of a codepoint into its payload, and moving
the functions for doing UTF-8 encoding into headers so they could be
inlined.

The work allowed the creation of some tests mixing BINARY! and CHAR!
and TEXT! in PARSE.