Unify ANY-STRING! and ANY-WORD! routines STR_XXX
During the transition to UTF-8 Everywhere, names like "UNI_AT()"
were used for dealing with unicode positioning in strings.  This
changes it so that they use the common name STR_XXX for their
routines.  However, they have different abilities...the word series
have symbol IDs and canons, while the string series have caches of
length and mappings of codepoint indices to offsets.

Also includes the ability to seek strings from either the head or the
tail, based on which is closer to the index being sought.