Add blockscope "fake keyword"
It can look confusing to write:

    {
        int x = 10;
        ...
        y = y + x;
    }

Sometimes that seems like it's unnecessary, and you're not quite sure
if it's getting picked up by some branch.  This adds a synonym for
NOOP:

    blockscope {
        int x = 10;
        ...
        y = y + x;
    }

It's better than a comment, uses the formal term for what it's doing,
and looks less like something you should just edit out.