First try at "bookmarking" strategy for UTF-8 Everywhere
There isn't O(1) access in the general case for UTF-8 series to seek
an index.  It can be if you know all the codepoints are ASCII, but
otherwise one has to walk character by character.

The initial basic implementation was done in such a way that it would
just seek from the beginning of the series and move character by
character.  For very long strings that were doing an iteration, this
was prohibitive--large parses such as the source analysis could run
all day and not finish.

This introduces a first take on a scheme that drastically improves
matters, by adding the idea of "bookmarks".  These are caches that
live with the series node, and are maintained accurately (or thrown
out) when edits are made to the string data.  They map from codepoints
to index positions.

While this is extremely basic and still relatively slow, it brings the
timing down more to earth, and allowing parsing to complete in minutes
instead of hours/days.  It uses some amount of intelligence to not
add caches to small series, and will walk from whichever is closest...
the bookmark or the head or tail of the string.