Panic in DEBUG_CHECK_CASTS macros SER()/etc.
This solidifies that the added checking on node integrity for
DEBUG_CHECK_CASTS is a C++11 only feature (would be too slow in
a C build and have to be rewritten, not worth it).  This does a
runtime check whenever the types being cast to are "more derived"
than the value in hand's type.

Also changes this to a panic instead of an assert for the failure.
This is because the builds which enable the ability are on Travis
and otherwise do not present a helpful point to debug.