Fix defective SER() usage for storing debug tick
The SER() helper adds an extra layer of checking in certain debug
builds at runtime that the thing being cast actually is a REBSER.  For
some reason it started catching a bad usage in Free_Node().  Since this
only runs in the occasional build on Travis, it showed up as an
assertion there.

The bad usage was related to debug-only information that may be
helpful.  So this just uses a plain cast.