Simplify parse, rename SET/GET_ANY_CHAR()
There was some optimization code in PARSE that complicated it a bit
by trying to notice when the first character of something matched or
not, and then not bothering to match the rest if it didn't.  This
optimization should be folded into the routine being called via
inlining, otherwise it winds up doubly seeking instead of being able
to use a single seek.

So that optimization should be looked at in a new light if it is
going to be done.  This gets rid of it, and the VAL_ANY_CHAR()
macro that it was using.  It also contextualizes the routines as
SET_CHAR_AT() and GET_CHAR_AT() and explains the narrow situations in
which they are useful.