Reorganize %sys-string.h and %sys-word.h
This improves the comments and moves the symbol-specific code into
%sys-word.h.  Increasingly it is appearing that the string and word
unificiation is going to basically stop at UTF-8 Everywhere, and that
they will be distinct classes...now that REFINEMENT! is gone, and
ISSUE! is slated to become a string, and SYMBOL-XXX! are on the agenda.
So starting to split things up makes sense.