TRANSCODE takes TEXT! as in/out, returns position, sets var
With UTF-8 Everywhere, it's now possible to run the scanner directly
on Rebol strings without converting them to BINARY! first.  Previous
tricks for transcoding TEXT! elements one at a time required keeping
the original text pointer, converting to binary, measuring the amount
of text that the consumed binary represented and adding it in as an
offset to the original text (!)

Besides being drastically more efficient than that, this also changes
the interface so that transcode takes a variable to set as a parameter
and returns the new position in the binary or text.  This makes it
align more closely with PARSE...as the ultimate goal is to replace
TRANSCODE the API entirely with PARSE.

The bootstrap shim integrates the old behavior of LOAD-NEXT as a
workaround, so that non-bootstrap builds will get the benefit of the
transcoding.