case-sensitive/case-insensitive string-in-binary find
We consider it always desirable to search for BINARY! in another
BINARY! with case-sensitive matching, e.g. to interpret the bytes not
as UTF-8 with a case-insensitive layering on it.

However, we offer a new ability of being able to match TEXT! as its
UTF-8 equivalent against a BINARY! (which may or may not be valid
UTF-8).  This matching can be done in either a case-sensitive or a
case-insensitive fashion.

This was intended when trying to rewrite the find routines, but wasn't
being exposed correctly in PARSE.  Fixes that, and adds some more
UTF-8 in PARSE tests.