Beginnings of PARSE of TEXT! being able to SET ANY-VALUE!
This is a very early test of being able to use PARSE to perform a
TRANSCODE-like operation.

    parse "123 [a b c] {def}" [
        set i integer!
        set b block!
        set t text!
        end
    ]

The semantics are unclear about how this will fit into parse rules
generally, e.g. to use with COPY...where it would be ambiguous whether
you wanted to copy values or the range of the string covered by
the parsing.

This is being committed since more work builds on it, and it should
not break anything--but due to its highly experimental status it should
not be relied upon (e.g. the operation that does this may not be
called SET).