Clarify message for `do [x: ()]`, `do [x: comment "hi"]`
These situations are distinct from `do [x: print "Hi"]`, because no
value is being produced.  But previously they would all give the
same error, claiming that a VOID! could not be assigned to x.

This is an improvement over the previous error situation, which says
that "an end was reached during evaluation".  That's still a little
bit misleading since there may be more values to come, but it's a
kind of virtual end.  (like how | creates an "end")

The main thing is just to cognitively assist people in knowing that
there is no VOID! involved.