Add SEEK and MARK to PARSE, disallow switching series
The notation of using SETWORD: to set a variable from the parse
position and :GETWORD to set the parse position from the variable is
a bit confusing.  (Early on, I'd frequently get confused what's being
set or gotten...)

Besides clarity, there is a good reason to have MARK and SEEK keywords.
This allows the use of expressions, such as seeking something that
was calculated.  Such as `mark @(get-current-var mode)`, or to use
a calculated position like `seek @(skip position length of thing)`

(That notation is based on the speculative concept that new SYMBOL
GROUP!s will be retaking the current role of the SET-GROUP!)

So this adds MARK and SEEK (though a simplified form that only allows
taking WORD! and SET-WORD! directly, for now).  The code path is shared
with SET-WORD! and GET-WORD! from historical PARSE.

Removes a little-known property of TO which let you seek TO an
INTEGER! position (which might be behind the current parse position).
SEEK takes care of that more clearly.

Also reverses an R3-Alpha decision to allow switching the source series
when seeking.  The ability hadn't been vetted completely for its
impacts...but the incremental parsing motivations for it are now
better handled by letting the parse return how far it has gotten and
starting a new parse.  Also, being able to switch the series would
throw off a number of features planned for unifying PARSE with the
TRANSCODE function.