Make type hook table bundle hooks consecutively in memory
Previously there were separately generated tables for each of the
type behavior hooks (comparison, molding, "generics", etc).  This
unifies them all into one table, but one that is arranged such that
its dispatchers for a single type are all consecutive.  This makes
it possible to pass around such a table entry via a single pointer--
e.g. one that could be poked into the EXTRA() field of a cell to
implement an extension type.