reb.Run() => reb.Value(), add reb.V()
The name rebRun() does not very clearly convey that a value will be
returned, making it easy to forget and just use it for something like
`rebRun("print {Hello}");`.  This generates an API handle for holding
the VOID! value cell, which if it isn't assigned to a value will leak.

This changes the API to be named rebValue(), and also adds the reb.V()
helper...for doing a kind of impromptu COMPOSE into an executing
stream of API code.