Move GC Debug Checks to separate %d-gc.c file
Because it contained a switch() on every datatype and had to take cells
apart in order to know how to GC mark them, the GC was a good place to
put in checks on the internal structure of cells.  So there were many
asserts and invariants checked.

This meant that finding corruptions could be done fairly effectively
by running a Recycle() on every operation, e.g. with RECYCLE/TORTURE.

New ideas for the GC that make it more extensible involve having the
cells carrying bits that say whether nodes in them need marking, and
Ren-C has a header system by which nodes can be differentiated from
each other.  This allows the GC to be more generic and not need a
switch statement.

This commit splits out that switch() statement for debugging, phrased
as a "check that the marking done without a switch worked just as
well as one that had one did".  This is an interim situation, and the
routine will likely become a more basic Assert_Cell() when it is
taken for granted that the marking worked correctly.