Load worker.js from Travis location vs. locally
The build uses a workaround that works for hosted libraries, so this
updates the loader to use the hosted location (instead of expecting
the host to have libr3.worker.js).

Also gets rid of the misleading non-emscripten-related "worker.js"
(which was just an attempt to salvage some code for creating workers
that had been commented, but was not needed once emscripten pthreads
were being used).

Deletes the committed libr3.worker.js file