Move libr3.js patch script to %make/tools/
The `clean` step of make would remove `libr3*`, and that was part of
the filename of the tool script for patching libr3 to use fetch()
and a Blob to load workers instead of `new Worker()`.

This moves it to the tools directory so it doesn't get deleted.

Takes care of an issue where there are two find strings, and we want
to substitute the second.

Doesn't try to do the patch on the emterpreter build (it does not
include the Pthread emulation code, so no `new Worker` to replace)