Reorganize, add LICENSE.txt for JavaScript Extension
Using GNU LGPL v3.0:

https://www.gnu.org/licenses/lgpl-3.0.html

Files and content from the Affero GPL-licensed ReplPad are being
incorporated into the JavaScript extension.  It appears there will be
a lot of back-and-forth of that; having different licenses is not
convenient.  So both are just going to be LGPL.

(Note: %mod-javascript.c had inherited "Apache 2.0" text from its
comment boilerplate.  As the sole author of this code--responsible for
whatever is good and bad about it--I authorize the change to LGPL.)