Use pairing instead of singular for quoted's cell node
This changes the QUOTED! cell type to use an ordinary node payload.

The current concept for how this is marked if it is a cell pointer is
to think of it as a pairing.  It would be possible to use something
like an API value and have it mark the containing array...which it will
probably be necessary to do anyway (e.g. generalized node marking for
extension REBSER or REBVAL very likely want API values to be a legal
choice for that pointer).  But since PAIR! uses a pairing, this goes
ahead and uses that for the time being.