Fold ARRAY_HAS_FLAG_FILE_LINE marking into generic REBSER mark
REBSER nodes are migrating to a system where there is a SERIES_FLAG_XXX
for whether each of the .link and .misc fields need to be marked.
With more node types knowing what kind of node they are, it can be
done this way instead of having switch() statements on each cell type.
The ultimate goal is to have a much smaller GC loop.

The previous special handling for marking ARRAY_FLAG_HAS_FILE_LINE is
changed here to where ARRAY_MASK_HAS_FILE_LINE is used, which carries
both the bit for ARRAY_FLAG_FILE_LINE_UNMASKED and the bit saying that
the link needs marking.