Use generalized node for word spelling synonym
This changes the REBSTR to use the link.common.node for its synonym.
However, it does not need to mark it--because it actually wants an
unused synonym to GC.  The linkages will be updated at the moment of
deletion.

While the marking is not used, what this paves the way for is for the
word or non-word status to be readable based on the string flag status
of what's in the link field.  This brings another step in unification,
so that word series can be read-only aliased as strings, without
having their synonym lists overwritten by bookmarks.