Get rid of Snapshot_All_Actions() experiment
Prior to the current plan of generalized evaluator hooking, there
was an attempt to snapshot a list of all the functions in the system
and inject timing information into them.  This was an attempt to give
meaning to:

    for-each a action! [...]

Today's concept of FOR-EACH is to lock the collection it is enumerating
and that's not very good if it's a list of all actions in the system.
But more generally, this is a questionable idea...which may expose
things that shouldn't be seen.

So at best, it would be a debug feature.  It hasn't been used and is
just taking up space.  So this removes the experiment.