Allow READ data to accumulate for SYNC-OP loop
While beginning the transition of the REBREQ code away from a structure
that the GC had to understand in particular to a generic REBSER, the
request was made to process BINARY! REBVALs directly instead of byte
pointers.  This meant the device layer was responsible for updating
the length of the binary series, vs. making that the responsibility
of the "higher level" ports.  (Eventually these are intended to unify
and be the same codebase, both using libRebol).

It wasn't exactly clear what the rules for READ were--in terms of
whether each READ operation was supposed to start fresh or if it could
leave data in the buffer.  An assert was added to check, and it did
not seem to fire--every call to READ seemed to be reusing the buffer
from a clean slate.

In practice this turned out to not be the case, as the SYNC-OP loop in
%prot-http expected to be able to keep adding and then copy the data
out as a whole chunk.  It's not clear why this wasn't previously
causing the assert to fire.  This documents the new finding and
includes a few improvements to %prot-http.r made while hunting the
bug down.