Switch to generic LINK() and MISC() GC marking
Previously there was specific knowledge based on the ARRAY_FLAG saying
if an array was a PARAMLIST vs. a VARLIST to decide how to mark its
LINK() and MISC() fields.  This takes another step deeper into the
generalization of that, by using the SERIES_FLAG_LINK_NODE_NEEDS_MARK
and SERIES_FLAG_MISC_NODE_NEEDS_MARK to guide that process.  This
further reduces the complexity of the release build GC code.