Change to SERIES_MASK_VARLIST, SERIES_MASK_PARAMLIST
Initially, the idea was to use a name to differentiate the mask from
the flag in a more obviously visible way...so SERIES_MASK_CONTEXT
was more obviously different from SERIES_FLAG_CONTEXT.  This creates
confusion with the likes of CELL_MASK_CONTEXT, however.

So since the flags have been named ARRAY_FLAG_IS_VARLIST and are
generally used via `GET_ARRAY_FLAG(a, IS_VARLIST)`, it makes more sense
to change to use the term for which piece of the context and action
it actually is.