Correct the ELSE usage in DO of a TAG!
ELSE completes only one expression on the left.  It was being used in
a way that suggested it was pairing with a switch, when it was
actually pairing with the clause that was being used as the switch
body.  Putting that clause in a GROUP! helps.

(This may be an argument for why SWITCH should soft-quote its body.)