Add special looking F4EE-trash and 5AFE-trash
When a series node is freed, it's good to have a special form of
trash with a pattern in it to show that's what happened.  (It's easier
to see than the series header).  This uses 0xF4EEF4EE instead of the
more generic 0xDECAFBAD.

It also after allocation uses 0x5AFE5AFE indicating it's trash, but
that it's safe to overwrite it and use it for the node.