Workaround gcc 4.6.3 warning
Trying to make the Manage_Array() variants of Manage_Series() return
an array for chaining caused:

../src/include/sys-action.h: In function ‘Init_Action_Unbound’:
../src/include/sys-action.h:439:5: error: value computed is not used
../src/include/sys-action.h: In function ‘Init_Action_Maybe_Bound’:
../src/include/sys-action.h:453:5: error: value computed is not used
In file included from ../src/include/sys-core.h:527:0,
                 from ../src/core/a-globals.c:33:

It would be possible to do -Wno-error-unused on this build, but there
were all of two uses of the chaining.  So this just gets rid of it.