%extensions/ to top level dir, put JS prep in JS extension
This takes the portion of the libRebol prep code that is needed for
JavaScript and factors it out into a file that lives in the JavaScript
extension.

It also moves %extensions/ to be a top level directory, and then takes
the necessary %external/ dependencies and puts them in the extensions
that own them.  This makes sense, as Rebol core seeks to be independent
and not need any third party repositories to build.