Series assertion-related fixes
There was a transitional UTF-8 Everywhere assert which overrode the
SER_WIDE() check in SER_AT_RAW().  That's no longer needed, because
the series is actually the correct byte width (1 for bytes, not a
fixed codepoint size).

This required fixing a hashing call that was claiming the size of the
series was 2 (the real size was used).  In fixing the call, this also
took BINARY! out of the category of things using a case insensitive
hash, since it doesn't do case-insensitive lookups.