Fix pairing GC problem in DEBUG_TRACK_EXTEND_CELLS
There are 3 forms of structure that can be in the SER_POOL; a regular
series with an allocation, a non-dynamic series with data in the node,
or two REBVALs.  It is plotted out so the sizes are the same.

Some compilation modes (and perhaps needed for exotic architectures)
allow this to be broken, where the two-REBVALs have to live in a
different pool.  This is UNUSUAL_REBVAL_SIZE, and it's used by a debug
mode called DEBUG_TRACK_EXTEND_CELLS which doubles the size of a REBVAL
in order to let it carry along extra information.

This mode had a bug in it that had not been noticed, related to the
GC of elements inside the series.  It hadn't been noticed because
previously the two-REBVAL-node mode had only been used to store items
like INTEGER! and DECIMAL for PAIR! which didn't participate in GC.

In addition to fixing that bug (by enumerating whole segments in
the pool instead of just their first item), this pushes the strict
aliasing compliance a bit further, by doing discernment of unknown
node types through a byte pointer instead of the headers of structures
that may be different (REBVAL->header vs. REBSER->header)