Make Is_Bindable() test depend on MIRROR_BYTE()
There had to be a previous exception made for types greater than
REB_MAX to say they were not bindable, which slowed the test via
another comparison.  Now that there are mirror bytes, the test can
be based on that value and leave the KIND_BYTE() as-is.

This means a REB_P_XXX parameter class declares its mirror byte to
actually be a REB_TYPESET, so it is more like a container for that.

The Is_Bindable() test is sped up as a result.