Fix strict equality compare for dates w/no time zone
This incorporates a couple of tests from rebol-issues bugs being
closed as fixed in R3-Alpha, so that the code from the bug is preserved

Doing so noticed that strict equality comparisons weren't seeing
synthetic dates made from the /DATE refinement on the result of NOW
as being equal to literal dates with no time zone.  This was because
the NO_DATE_ZONE field wasn't being set when NO_DATE_TIME was set.

Works around that issue by canonizing the rule that dates with
NO_DATE_TIME also have NO_DATE_ZONE.