Fix CHANGE length issue
When CHANGE of an ANY-STRING! was used without /PART it was replacing
to the end of the series.  This makes it just replace the amount of
material given.

There were almost no tests of CHANGE (hence such problems would easily
go unnoticed--so tests should be gathered or written).  But this adds a
basic test for this specific behavior.