Tweak to try and handle CHANGE with different codepoint sizes
With UTF-8 Everywhere, the behavior of CHANGE gets more complicated
because what had previously appeared to not require series expansion
(e.g. a CHANGE of one character) now can be inserting codepoints of
the same length, but different size.

In lieu of having time to do a complete overhaul and vetting of the
code for CHANGE, this tries to patch it at least so that it can
handle a case like:

    >> s: "abc"
    == "abc"

    >> change s "ò"
    == "bc"

    >> s
    == "òbc"