Implicate line causing missing delimiter error
At some point, the error message's property telling you what line a
string scan failed on started getting overwritten, so it wouldn't tell
you the file and line that had the problem.  This corrects that by
not overwriting the information (which was beeing put on the error,
and just wiped out later.)

In addition, this updates string scans to report a more accurate error
of the line where the string is...vs. the block in which the string
is being scanned.

Also avoids a bad error message when big, but valid codepoints are used
with the caret notation (pending a better error message in that code)